chore: rename dump-sources to print-sources
