+++ tests/integration/_cases/debug_files/debug_files-no-subcommand.trycmd
@@ -21,9 +21,9 @@ OPTIONS:
 SUBCOMMANDS:
     bundle-sources    Create a source bundle for a given debug information file
     check             Check the debug info file at a given path.
-    dump-sources      Print source files linked by the given debug info file.
     find              Locate debug information files for given debug identifiers.
     help              Print this message or the help of the given subcommand(s)
+    print-sources     Print source files linked by the given debug info file.
     upload            Upload debugging information files.
 
 ```
